check the genome file

+++ main.nf
@@ -27,7 +27,7 @@ include {snp_footprint_clustering} from './modules/snp_footprint_clustering'
 process downloadGenome {
 
     label 'low_cpu_low_mem'
-    tag "Sample - $sampleId" 
+    tag "Dowloading - $url" 
     publishDir "${projectDir}/ref_files/genome", mode : 'copy'
 
     container = "quay.io/biocontainers/wget:1.21.4"
@@ -41,8 +41,12 @@ process downloadGenome {
     script:
     def url = genome == 'hg19' ? params.hg19GenomeDownload : params.hg38GenomeDownload
     """
-    wget -O genome.fa.gz ${url}
-    gunzip genome.fa.gz
+    if [ ! -f genome.fa ]; then
+        wget -O genome.fa.gz ${url}
+        gunzip genome.fa.gz
+    else
+        echo "File genome.fa already exists. Skipping download."
+    fi
     """
 }
